Move toml logic to separate class ### What's done: * Move toml logic to separate class
